move generic test to test_generics.py
